machine/requirements.txt
asgiref==3.4.1
Django==3.2.25
fasttext-wheel==0.9.2
numpy==1.19.5
pybind11==2.12.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.1.1
